Move all inc_encode and inc_reconstruct function to ReedSolomon struct
